Moved link into media tab
